7/13
Worked on root find function.  It seems to work for some winding
numbers and not others.  Not sure why at this point.
Also need to fix the recession depth problem.  I think it is doable by
making a list of functions and applying them in order… might not work
with scipi.optimize.